requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.4
sse-starlette==2.2.0
dspy-ai==2.5.13
openai==1.58.1
anthropic==0.40.0
google-generativeai==0.8.3
httpx==0.28.1
python-multipart==0.0.12
cors==1.0.1 